Move frame output structs to dedicated header
