+++ src/torchcodec/_core/CudaDevice.cpp
@@ -192,7 +192,7 @@ void CudaDevice::initializeContext(AVCodecContext* codecContext) {
 void CudaDevice::convertAVFrameToFrameOutput(
     const VideoStreamOptions& videoStreamOptions,
     UniqueAVFrame& avFrame,
-    SingleStreamDecoder::FrameOutput& frameOutput,
+    FrameOutput& frameOutput,
     std::optional<torch::Tensor> preAllocatedOutputTensor) {
   TORCH_CHECK(
       avFrame->format == AV_PIX_FMT_CUDA,
