+++ src/torchcodec/_core/SingleStreamDecoder.h
@@ -14,7 +14,8 @@
 
 #include "src/torchcodec/_core/AVIOContextHolder.h"
 #include "src/torchcodec/_core/FFMPEGCommon.h"
-#include "src/torchcodec/_core/StreamOptions.h"
+#include "src/torchcodec/_core/Frame.h"
+#include "src/torchcodec/_core/Stream.h"
 
 namespace facebook::torchcodec {
 class DeviceInterface;
@@ -52,56 +53,6 @@ class SingleStreamDecoder {
   // the allFrames and keyFrames vectors.
   void scanFileAndUpdateMetadataAndIndex();
 
-  struct StreamMetadata {
-    // Common (video and audio) fields derived from the AVStream.
-    int streamIndex;
-    // See this link for what various values are available:
-    // https://ffmpeg.org/doxygen/trunk/group__lavu__misc.html#ga9a84bba4713dfced21a1a56163be1f48
-    AVMediaType mediaType;
-    std::optional<AVCodecID> codecId;
-    std::optional<std::string> codecName;
-    std::optional<double> durationSeconds;
-    std::optional<double> beginStreamFromHeader;
-    std::optional<int64_t> numFrames;
-    std::optional<int64_t> numKeyFrames;
-    std::optional<double> averageFps;
-    std::optional<double> bitRate;
-
-    // More accurate duration, obtained by scanning the file.
-    // These presentation timestamps are in time base.
-    std::optional<int64_t> minPtsFromScan;
-    std::optional<int64_t> maxPtsFromScan;
-    // These presentation timestamps are in seconds.
-    std::optional<double> minPtsSecondsFromScan;
-    std::optional<double> maxPtsSecondsFromScan;
-    // This can be useful for index-based seeking.
-    std::optional<int64_t> numFramesFromScan;
-
-    // Video-only fields derived from the AVCodecContext.
-    std::optional<int64_t> width;
-    std::optional<int64_t> height;
-
-    // Audio-only fields
-    std::optional<int64_t> sampleRate;
-    std::optional<int64_t> numChannels;
-    std::optional<std::string> sampleFormat;
-  };
-
-  struct ContainerMetadata {
-    std::vector<StreamMetadata> allStreamMetadata;
-    int numAudioStreams = 0;
-    int numVideoStreams = 0;
-    // Note that this is the container-level duration, which is usually the max
-    // of all stream durations available in the container.
-    std::optional<double> durationSeconds;
-    // Total BitRate level information at the container level in bit/s
-    std::optional<double> bitRate;
-    // If set, this is the index to the default audio stream.
-    std::optional<int> bestAudioStreamIndex;
-    // If set, this is the index to the default video stream.
-    std::optional<int> bestVideoStreamIndex;
-  };
-
   // Returns the metadata for the container.
   ContainerMetadata getContainerMetadata() const;
 
@@ -123,38 +74,6 @@ class SingleStreamDecoder {
   // --------------------------------------------------------------------------
   // DECODING AND SEEKING APIs
   // --------------------------------------------------------------------------
-
-  // All public video decoding entry points return either a FrameOutput or a
-  // FrameBatchOutput.
-  // They are the equivalent of the user-facing Frame and FrameBatch classes in
-  // Python. They contain RGB decoded frames along with some associated data
-  // like PTS and duration.
-  // FrameOutput is also relevant for audio decoding, typically as the output of
-  // getNextFrame(), or as a temporary output variable.
-  struct FrameOutput {
-    // data shape is:
-    // - 3D (C, H, W) or (H, W, C) for videos
-    // - 2D (numChannels, numSamples) for audio
-    torch::Tensor data;
-    double ptsSeconds;
-    double durationSeconds;
-  };
-
-  struct FrameBatchOutput {
-    torch::Tensor data; // 4D: of shape NCHW or NHWC.
-    torch::Tensor ptsSeconds; // 1D of shape (N,)
-    torch::Tensor durationSeconds; // 1D of shape (N,)
-
-    explicit FrameBatchOutput(
-        int64_t numFrames,
-        const VideoStreamOptions& videoStreamOptions,
-        const StreamMetadata& streamMetadata);
-  };
-
-  struct AudioFramesOutput {
-    torch::Tensor data; // shape is (numChannels, numSamples)
-    double ptsSeconds;
-  };
 
   // Places the cursor at the first frame on or after the position in seconds.
   // Calling getNextFrame() will return the first frame at
@@ -536,7 +455,7 @@ FrameDims getHeightAndWidthFromResizedAVFrame(const AVFrame& resizedAVFrame);
 
 FrameDims getHeightAndWidthFromOptionsOrMetadata(
     const VideoStreamOptions& videoStreamOptions,
-    const SingleStreamDecoder::StreamMetadata& streamMetadata);
+    const StreamMetadata& streamMetadata);
 
 FrameDims getHeightAndWidthFromOptionsOrAVFrame(
     const VideoStreamOptions& videoStreamOptions,
